Boston Housing Data Regression using Deep Learning
Prediction of housing prices using deep learning, as opposed to classical machine learning methods that are usually employed for this small dataset. The notebook attempts to exemplify techniques and methods that can be employed on these complex models to reduce the effects of over-fitting, along with showing an example of how a neural network can be used for a regression task.